requirements.txt
Flask
Flask-Login
Flask-Migrate
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Werkzeug
SQLAlchemy
pandas
openpyxl
fpdf2
gunicorn